add doccomments

+++ v1/card.go
@@ -2,29 +2,31 @@ package payjp
 
 import (
 	"encoding/json"
+	"errors"
 	"time"
 )
 
+// Card はCustomerやTokenのAPIでカード情報を設定する時に使う構造体です
 type Card struct {
-	Number       interface{}
-	ExpMonth     int
-	ExpYear      int
-	CVC          int
-	AddressState interface{}
-	AddressCity  interface{}
-	AddressLine1 interface{}
-	AddressLine2 interface{}
-	AddressZip   interface{}
-	Country      interface{}
-	Name         interface{}
+	Name         interface{} // カード保有者名(e.g. YUI ARAGAKI)
+	Number       interface{} // カード番号
+	ExpMonth     int         // 有効期限月
+	ExpYear      int         // 有効期限年
+	CVC          int         // CVCコード
+	Country      interface{} // 2桁のISOコード(e.g. JP)
+	AddressZip   interface{} // 郵便番号
+	AddressState interface{} // 都道府県
+	AddressCity  interface{} // 市区町村
+	AddressLine1 interface{} // 番地など
+	AddressLine2 interface{} // 建物名など
 }
 
-func (c Card) Valid() bool {
+func (c Card) valid() bool {
 	_, ok := c.Number.(string)
 	return ok && c.ExpYear > 0 && c.ExpMonth > 0
 }
 
-func (c Card) Empty() bool {
+func (c Card) empty() bool {
 	_, ok := c.Number.(string)
 	return !ok && c.ExpYear == 0 && c.ExpMonth == 0
 }
@@ -39,23 +41,24 @@ func parseCard(service *Service, body []byte, result *CardResponse, customerID s
 	return result, nil
 }
 
+// CardResponse はCustomerやTokenのAPIが返す構造体です
 type CardResponse struct {
-	AddressCity     string
-	AddressLine1    string
-	AddressLine2    string
-	AddressState    string
-	AddressZip      string
-	AddressZipCheck string
-	Brand           string
-	Country         string
-	CreatedAt       time.Time
-	CvcCheck        string
-	ExpMonth        int
-	ExpYear         int
-	Fingerprint     string
-	ID              string
-	Last4           string
-	Name            string
+	CreatedAt       time.Time // カード作成時のタイムスタンプ
+	ID              string    // car_で始まる一意なオブジェクトを示す文字列
+	Name            string    // カード保有者名(e.g. YUI ARAGAKI)
+	Last4           string    // カード番号の下四桁
+	ExpMonth        int       // 有効期限月
+	ExpYear         int       // 有効期限年
+	Brand           string    // カードブランド名(e.g. Visa)
+	CvcCheck        string    // CVCコードチェックの結果
+	Fingerprint     string    // このクレジットカード番号に紐づけられた一意（他と重複しない）キー
+	Country         string    // 2桁のISOコード(e.g. JP)
+	AddressZip      string    // 郵便番号
+	AddressZipCheck string    // 郵便番号存在チェックの結果
+	AddressState    string    // 都道府県
+	AddressCity     string    // 市区町村
+	AddressLine1    string    // 番地など
+	AddressLine2    string    // 建物名など
 
 	customerID string
 	service    *Service
@@ -81,15 +84,26 @@ type cardResponseParser struct {
 	Object          string `json:"object"`
 }
 
+// Update メソッドはカードの内容を更新します
+// Customer情報から得られるカードでしか更新はできません
 func (c *CardResponse) Update(card Card) error {
+	if c.customerID == "" {
+		return errors.New("Token's card doens't support Update()")
+	}
 	_, err := c.service.Customer.postCard(c.customerID, "/"+c.ID, card, c)
 	return err
 }
 
+// Delete メソッドは顧客に登録されているカードを削除します
+// Customer情報から得られるカードでしか削除はできません
 func (c *CardResponse) Delete() error {
+	if c.customerID == "" {
+		return errors.New("Token's card doens't support Delete()")
+	}
 	return c.service.delete("/customers/" + c.customerID + "/cards/" + c.ID)
 }
 
+// UnmarshalJSON はJSONパース用の内部APIです。
 func (c *CardResponse) UnmarshalJSON(b []byte) error {
 	raw := cardResponseParser{}
 	err := json.Unmarshal(b, &raw)
